Refactor yaml save utility
